updates suggested by Chris

+++ Validation/Performance/python/TimeMemorySummary.py
@@ -8,7 +8,7 @@ def customise(process):
     #Adding Timing service:
     process.Timing=cms.Service("Timing",
                                summaryOnly=cms.untracked.bool(True),
-                               excessiveTimeThreshold=cms.untracked.double(100))
+                               excessiveTimeThreshold=cms.untracked.double(600))
     
     #Add these 3 lines to put back the summary for timing information at the end of the logfile
     #(needed for TimeReport report)
